formatting of tools utils
