+++ tests/test_run_train.py
@@ -326,27 +326,27 @@ def test_run_train_foundation(tmp_path, fitting_configs):
     print("Es", Es)
     # from a run on 28/03/2023 on repulsion a63434aaab70c84ee016e13e4aca8d57297a0f26
     ref_Es = [
-        1.6488910913467407,
-        0.79408198595047,
-        0.5990670919418335,
-        0.4938647150993347,
-        0.5145026445388794,
-        0.5850723385810852,
-        0.4688843786716461,
-        0.5712210536003113,
-        0.5261895060539246,
-        0.39647144079208374,
-        3.34824538230896,
-        0.5676671266555786,
-        0.5044677257537842,
-        0.7869258522987366,
-        0.6437031030654907,
-        0.548050045967102,
-        0.7052161693572998,
-        0.6819053888320923,
-        0.6876872181892395,
-        0.32130151987075806,
-        0.483511358499527,
-        0.597072422504425,
+        1.6324548721313477,
+        0.8016132712364197,
+        0.5674604177474976,
+        0.4571378827095032,
+        0.5002979040145874,
+        0.5423622131347656,
+        0.4228057265281677,
+        0.6198095083236694,
+        0.4928140640258789,
+        0.34875422716140747,
+        3.448421001434326,
+        0.5291348695755005,
+        0.463774710893631,
+        0.7975829243659973,
+        0.6095817685127258,
+        0.5113707780838013,
+        0.6703207492828369,
+        0.6445103883743286,
+        0.6734066605567932,
+        0.25074106454849243,
+        0.44250693917274475,
+        0.5659750699996948,
     ]
     assert np.allclose(Es, ref_Es)
